+++ src/Serilog.Sinks.Udp/LoggerSinkConfigurationExtensions.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2015-2016 Serilog Contributors
+﻿// Copyright 2015-2017 Serilog Contributors
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,23 +13,49 @@
 // limitations under the License.
 
 using System;
+using System.ComponentModel;
 using System.Net;
 using Serilog.Configuration;
 using Serilog.Events;
+using Serilog.Formatting;
 using Serilog.Formatting.Display;
-using Serilog.Sinks.Udp;
+using Serilog.Sinks.Udp.Private;
 
 namespace Serilog
 {
     /// <summary>
-    /// Extends <see cref="LoggerConfiguration"/> to add UDP sink capabilities.
+    /// Adds the WriteTo.Udp() extension method to <see cref="LoggerConfiguration"/>.
     /// </summary>
     public static class LoggerSinkConfigurationExtensions
     {
-        private const string DefaultOutputTemplate = "{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz} [{Level}] {Message}{NewLine}{Exception}";
+        private const string DefaultOutputTemplate = "{Timestamp:yyyy-MM-dd HH:mm:ss} [{Level}] {Message}{NewLine}{Exception}";
 
         /// <summary>
-        /// Send log events as UDP packages over the network.
+        /// Extension method providing JSON support via
+        /// <a href="https://github.com/serilog/serilog-settings-configuration">Serilog.Settings.Configuration</a>.
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public static LoggerConfiguration Udp(
+            this LoggerSinkConfiguration sinkConfiguration,
+            string remoteAddress,
+            int remotePort,
+            int localPort = 0,
+            LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum,
+            string outputTemplate = DefaultOutputTemplate,
+            IFormatProvider formatProvider = null)
+        {
+            return Udp(
+                sinkConfiguration,
+                IPAddress.Parse(remoteAddress),
+                remotePort,
+                localPort,
+                restrictedToMinimumLevel,
+                outputTemplate,
+                formatProvider);
+        }
+        
+        /// <summary>
+        /// Adds a sink that sends log events as UDP packages over the network.
         /// </summary>
         /// <param name="sinkConfiguration">Logger sink configuration.</param>
         /// <param name="remoteAddress">
@@ -50,12 +76,12 @@ namespace Serilog
         /// </param>
         /// <param name="outputTemplate">
         /// A message template describing the format used to write to the sink. The default is
-        /// "{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz} [{Level}] {Message}{NewLine}{Exception}".
+        /// "{Timestamp:yyyy-MM-dd HH:mm:ss} [{Level}] {Message}{NewLine}{Exception}".
         /// </param>
         /// <param name="formatProvider">
         /// Supplies culture-specific formatting information, or null.
         /// </param>
-        /// <returns>Configuration object allowing method chaining.</returns>
+        /// <returns>Logger configuration, allowing configuration to continue.</returns>
         public static LoggerConfiguration Udp(
             this LoggerSinkConfiguration sinkConfiguration,
             IPAddress remoteAddress,
@@ -65,10 +91,81 @@ namespace Serilog
             string outputTemplate = DefaultOutputTemplate,
             IFormatProvider formatProvider = null)
         {
-            if (sinkConfiguration == null) throw new ArgumentNullException(nameof(sinkConfiguration));
+            if (sinkConfiguration == null)
+                throw new ArgumentNullException(nameof(sinkConfiguration));
+            if (outputTemplate == null)
+                throw new ArgumentNullException(nameof(outputTemplate));
 
             var formatter = new MessageTemplateTextFormatter(outputTemplate, formatProvider);
-            var sink = new UdpSink(localPort, remoteAddress, remotePort, formatter);
+            return Udp(
+                sinkConfiguration,
+                remoteAddress,
+                remotePort,
+                formatter,
+                localPort,
+                restrictedToMinimumLevel);
+        }
+
+        /// <summary>
+        /// Extension method providing JSON support via
+        /// <a href="https://github.com/serilog/serilog-settings-configuration">Serilog.Settings.Configuration</a>.
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public static LoggerConfiguration Udp(
+            this LoggerSinkConfiguration sinkConfiguration,
+            string remoteAddress,
+            int remotePort,
+            ITextFormatter formatter,
+            int localPort = 0,
+            LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum)
+        {
+            return Udp(
+                sinkConfiguration,
+                IPAddress.Parse(remoteAddress),
+                remotePort,
+                formatter,
+                localPort,
+                restrictedToMinimumLevel);
+        }
+
+        /// <summary>
+        /// Adds a sink that sends log events as UDP packages over the network.
+        /// </summary>
+        /// <param name="sinkConfiguration">Logger sink configuration.</param>
+        /// <param name="remoteAddress">
+        /// The <see cref="IPAddress"/> of the remote host or multicast group to which the UDP
+        /// client should sent the logging event.
+        /// </param>
+        /// <param name="remotePort">
+        /// The TCP port of the remote host or multicast group to which the UDP client should sent
+        /// the logging event.
+        /// </param>
+        /// <param name="formatter">
+        /// Controls the rendering of log events into text, for example to log JSON. To control
+        /// plain text formatting, use the overload that accepts an output template.
+        /// </param>
+        /// <param name="localPort">
+        /// The TCP port from which the UDP client will communicate. The default is 0 and will
+        /// cause the UDP client not to bind to a local port.
+        /// </param>
+        /// <param name="restrictedToMinimumLevel">
+        /// The minimum level for events passed through the sink. The default is
+        /// <see cref="LevelAlias.Minimum"/>.
+        /// </param>
+        /// <returns>Logger configuration, allowing configuration to continue.</returns>
+        public static LoggerConfiguration Udp(
+            this LoggerSinkConfiguration sinkConfiguration,
+            IPAddress remoteAddress,
+            int remotePort,
+            ITextFormatter formatter,
+            int localPort = 0,
+            LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum)
+        {
+            if (sinkConfiguration == null)
+                throw new ArgumentNullException(nameof(sinkConfiguration));
+
+            var client = UdpClientFactory.Create(localPort, remoteAddress);
+            var sink = new UdpSink(client, remoteAddress, remotePort, formatter);
             return sinkConfiguration.Sink(sink, restrictedToMinimumLevel);
         }
     }