+++ src/Serilog.Sinks.Udp/Properties/AssemblyInfo.cs
@@ -1,22 +1,10 @@
-﻿using System.Reflection;
+﻿using System.Runtime.CompilerServices;
 
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("Serilog.Sinks.Udp")]
-[assembly: AssemblyDescription("Serilog sink for UDP")]
-[assembly: AssemblyCopyright("Copyright © Serilog Contributors 2015-2016")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.1.0")]
-[assembly: AssemblyFileVersion("1.0.1.0")]
-[assembly: AssemblyInformationalVersion("1.0.1")]
+[assembly: InternalsVisibleTo(
+    "DynamicProxyGenAssembly2, " +
+    "PublicKey=" +
+    "0024000004800000940000000602000000240000525341310004000001000100c547cac37abd99" +
+    "c8db225ef2f6c8a3602f3b3606cc9891605d02baa56104f4cfc0734aa39b93bf7852f7d9266654" +
+    "753cc297e7d2edfe0bac1cdcf9f717241550e0a7b191195b7667bb4f64bcb8e2121380fd1d9d46" +
+    "ad2d92d2d15605093924cceaf74c4861eff62abf69b9291ed0a340e113be11e6a7d3113e92484c" +
+    "f7045cc7")]